control-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
PyYAML==6.0.2
GitPython==3.1.43
python-dotenv==1.0.1
docker==7.1.0
mcrcon==0.7.0
psutil==6.0.0
APScheduler==3.10.4
tenacity==9.0.0